refactor: 팔로워 팔로잉 모달을 공통적인 컴포넌트로 묶음
